Split build for Bluetooth functionality to compile on non-Linux systems
